fixed script and added working with venv

+++ service/routers.py
@@ -35,7 +35,7 @@ async def upload_files(checksum: str, fi: UploadFile = File(default=None)):
             os.remove(upload_path)
         raise ValueError(f"File {fi.filename} has an incorrect checksum or format")
 
-    # await updater()
+    await updater(fi.filename)
     return {f"message": f"All files were successfully uploaded"}
 
 
@@ -67,25 +67,33 @@ async def kill_services():
         except:
             continue
     # os.system(f"chmod +x {os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/stop_all_services.sh')}")
-    # subprocess.call(os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/stop_all_services.sh'))
+    # subprocess.call(os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/stop_all_services.sh'), shell=True)
 
 
-async def restore_old_project():
+async def restore_old_project(filename, venv=False):
     try:
         # Remove archive
         try:
-            os.remove(os.path.join(UPLOAD_DIR, 'software/radius_control_backend.zip'))
+            os.remove(os.path.join(UPLOAD_DIR, filename))
         except Exception as e:
             print(e)
-        # Remove directory of new project
+        # Remove directory of new project and venv
         try:
             shutil.rmtree(os.path.join(UPLOAD_DIR, 'software/radius_control_backend/'))
+            if venv:
+                shutil.rmtree(os.path.join(UPLOAD_DIR, 'python3.10/'))
         except Exception as e:
             print(e)
-        # Rename old directory to radius_control_backend
+        # Rename old directory to radius_control_backend and venv
         try:
             os.rename(os.path.join(UPLOAD_DIR, 'software/backup_radius_control_backend/'),
                       os.path.join(UPLOAD_DIR, 'software/radius_control_backend/'))
+        except Exception as e:
+            print(e)
+        try:
+            if venv:
+                os.rename(os.path.join(UPLOAD_DIR, 'backup_python3.10/'),
+                          os.path.join(UPLOAD_DIR, 'python3.10/'))
         except Exception as e:
             print(e)
         # Run old project
@@ -93,35 +101,42 @@ async def restore_old_project():
             os.system(
                 f"chmod +x {os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/start_all_services.sh')}")
             subprocess.call(
-                os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/start_all_services.sh'))
+                os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/start_all_services.sh'), shell=True)
         except Exception as e:
             print(e)
     except Exception as e:
         raise HTTPException("Failed restore old project:", e)
 
 
-async def updater():
+async def updater(filename):
     # Kill all project services
     await kill_services()
     try:
         os.rename(os.path.join(UPLOAD_DIR, 'software/radius_control_backend/'),
                   os.path.join(UPLOAD_DIR, 'software/backup_radius_control_backend/'))
     except Exception as e:
+        # Exception is always in renaming, but it works
         print(e)
 
     try:
         # Unzipping new
-        with zipfile.ZipFile(os.path.join(UPLOAD_DIR, 'radius_control_backend.zip'), 'r') as zip_ref:
+        with zipfile.ZipFile(os.path.join(UPLOAD_DIR, filename), 'r') as zip_ref:
             zip_ref.extractall(UPLOAD_DIR)
         shutil.move(os.path.join(UPLOAD_DIR, 'radius_control_backend'),
                     (os.path.join(f'{UPLOAD_DIR}', 'software/radius_control_backend')))
         try:
-            os.rename(os.path.join(UPLOAD_DIR, 'venv/'),
-                      os.path.join(UPLOAD_DIR, 'python3.10/'))
+            if os.path.exists(os.path.join(UPLOAD_DIR, 'venv/')):
+                os.rename(os.path.join(UPLOAD_DIR, 'python3.10/'),
+                          os.path.join(UPLOAD_DIR, 'backup_python3.10/'))
+                os.rename(os.path.join(UPLOAD_DIR, 'venv/'),
+                          os.path.join(UPLOAD_DIR, 'python3.10/'))
         except Exception as e:
             print(e)
     except Exception as e:
-        await restore_old_project()
+        venv = False
+        if os.path.exists(os.path.join(UPLOAD_DIR, 'backup_python3.10/')):
+            venv = True
+        await restore_old_project(filename, venv=venv)
         raise HTTPException("Failed extract new project:", e)
 
     try:
@@ -129,10 +144,16 @@ async def updater():
         os.system(
             f"chmod +x {os.path.join(UPLOAD_DIR, 'software/radius_control_backend/common/third_party/start_all_services.sh')}")
         subprocess.call(os.path.join(UPLOAD_DIR,
-                                     'software/radius_control_backend/common/third_party/start_all_services.sh'))
+                                     'software/radius_control_backend/common/third_party/start_all_services.sh'), shell=True)
     except Exception as e:
-        await restore_old_project()
+        venv = False
+        if os.path.exists(os.path.join(UPLOAD_DIR, 'backup_python3.10/')):
+            venv = True
+        await restore_old_project(filename, venv=venv)
         raise HTTPException("Failed to run new project:", e)
     # Remove archive and directory of old project
-    os.remove(os.path.join(UPLOAD_DIR, 'radius_control_backend.zip'))
+    os.remove(os.path.join(UPLOAD_DIR, filename))
     shutil.rmtree(os.path.join(UPLOAD_DIR, 'software/backup_radius_control_backend/'))
+    if os.path.exists(os.path.join(UPLOAD_DIR, 'backup_python3.10/')):
+        shutil.rmtree(os.path.join(UPLOAD_DIR, 'backup_python3.10/'))
+    os.system("sudo reboot")
